requirements.txt
pytest
selene>=2.0.0rc2
allure-pytest==2.13.3